Socket Updates + MQTT integration raspi

+++ python/imageclient.py
@@ -3,7 +3,29 @@ import struct
 import socket
 import io
 import time
+from Adafruit_IO import MQTTClient
 
+client = MQTTClient('raebelchristo', 'aio_NHyV71dLrRu5QBLqhs0vT364IHTS')
+
+newDataReceived = False
+
+def connected(client):
+    client.subscribe('enteringcar')
+
+def disconnected(client):
+    print("Lost connection to Adafruit")
+    try:
+        client.connect()
+    except Exception as e:
+        print(f"Error occured during reconnection: [{e.strerr}]")
+        exit()
+
+def message(client, feed_id, payload):
+    print(f"Adafruit sent: [{payload}]")
+    if payload==1:
+        newDataReceived=True
+    else:
+        newDataReceived=False
 
 def establish_connection(sock):
     c = 1
@@ -25,38 +47,40 @@ def establish_connection(sock):
 
 sock = socket.socket()
 connection = establish_connection(sock)
+client.connect()
+client.loop_background()
 
 while True:
-    try:
+    if(newDataReceived):
+        try:
+            camera = picamera.PiCamera()
+            camera.resolution = (900,900)
+            camera.vflip = True
+            camera.hflip = True
+            camera.start_preview()
+            time.sleep(2)
 
-        camera = picamera.PiCamera()
-        camera.resolution = (900,900)
-        camera.vflip = True
-        camera.hflip = True
-        camera.start_preview()
-        time.sleep(2)
+            stream = io.BytesIO()
 
-        stream = io.BytesIO()
+            for foo in camera.capture_continuous(stream, 'jpeg'):
+                connection.write(struct.pack('<L', stream.tell()))
+                connection.flush()
+                stream.seek(0)
+                connection.write(stream.read())
+                stream.seek(0)
+                stream.truncate()
+                print("Streamed an image")
+                time.sleep(0.5)
 
-        for foo in camera.capture_continuous(stream, 'jpeg'):
-            connection.write(struct.pack('<L', stream.tell()))
-            connection.flush()
-            stream.seek(0)
-            connection.write(stream.read())
-            stream.seek(0)
-            stream.truncate()
-            print("Streamed an image")
-            time.sleep(0.5)
+        except KeyboardInterrupt:
+            print("Terminating Camera Stream")
+            connection.write(struct.pack('<L', 0))
+            break
 
-    except KeyboardInterrupt:
-        print("Terminating Camera Stream")
-        connection.write(struct.pack('<L', 0))
-        break
-
-    except socket.error as e:
-        print(f"Lost connection with host [{e.strerror}]. Retrying connection in 2 seconds")
-        time.sleep(2)
-        connection = establish_connection(sock)
+        except socket.error as e:
+            print(f"Lost connection with host [{e.strerror}]. Retrying connection in 2 seconds")
+            time.sleep(2)
+            connection = establish_connection(sock)
 
 connection.close()
 sock.close()
